Creation of the computation algorithm (#38)
* Creation of my algorithm

Woot

* Updated files based on checkstyle rules

* More Checkstyle fixes

Deleted some wonky comments, fixed an if statement error

* MORE CHECKSTYLE

RAHHHHH

* Delete Prime Numbers/PrimeGenerator.java

* Delete src/primes.csv

* Edits to the PrimeSummation Calculation

Edited the code to be able to account for more than 2 digits in the summation. Working on fixing repeat values still [3,7] and [7,3] for example. Just committing this to the branch so you can take a look if you would like.

* Quick Change to Computation

Splitting the listed numbers on a comma and a space (", ") instead of just a space. Also threw an exception for any negative numbers besides -1 input.

* Add a large file using Git LFS

* Edited the code and such

Changed primes.txt to include all primes up to the int limit (I think, I cannot be sure it didn't miss one or something). Then updated The computation to account for this (got rid of the limit to allow for much greater input values). Also removed the debugging line I had.

* Looking for checkstlye error

* Checkstyle

* Fixed formatting for if/else

Chekcstyle requires the else to be on the same line as the right curly brace of the if statement like this: } else {

---------

Co-authored-by: Brian  Abbott <113200930+babbo65@users.noreply.github.com>